+++ client/src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.scss';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 import { HomeComponent, Sidebar } from './home-component';
-import { Migrator } from "./migrator";
+import { Migrator } from "./migrator/migrator";
 import ErrorPage from "./error-404";
 
 class Validator extends React.Component {
